Factor out new data_capture.jobs module.
